Update README.md
This is an electricity bill calculator using python.  I have made this project using the if-else loop while I was learning to use the if-else loop.  It calculates the bill of the consumer by first taking input of the consumer name and meter reading from the consumer. Then, depending upon the unit consumed by the consumer it calculates the amount by multiplying the unit by rate and this rate varies with the unit consumed by the consumer.